+++ ena_upload/json_parsing/characteristic.py
