+++ ena_upload/json_parsing/json_schemas/assay_schema.json
