+++ ena_upload/json_parsing/json_schemas/investigation_schema.json
