+++ ena_upload/json_parsing/json_schemas/material_attribute_schema.json
