+++ ena_upload/json_parsing/json_schemas/material_schema.json
