+++ ena_upload/json_parsing/json_schemas/process_parameter_value_schema.json
