+++ ena_upload/json_parsing/json_schemas/protocol_parameter_schema.json
