+++ ena_upload/json_parsing/json_schemas/protocol_schema.json
