+++ ena_upload/json_parsing/json_schemas/publication_schema.json
