+++ ena_upload/json_parsing/json_schemas/study_schema.json
